move files so it does not conflict with mnist gitignore
