--- tests/datasets/__init__.py
@@ -1,1 +1,0 @@
-from tests.datasets.pl_mnist import MNIST
