--- tests/datasets/pl_mnist/__init__.py
@@ -1,1 +1,0 @@
-from tests.datasets.pl_mnist.mnist import MNIST
