+++ tests/pl_mnist/__init__.py
@@ -1,0 +1,2 @@
+from tests.pl_mnist.mnist import MNIST
+from tests.pl_mnist.transforms import *
