+++ tests/pl_mnist/utils.py
